:recycle: refactoring card list
